refactor: rename referralUtils to utils
