+++ src/commands/referral/referralNew.test.ts
@@ -4,10 +4,10 @@ import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { captor, mockDeep, mockReset } from 'vitest-mock-extended';
 import { getOrCreateUser } from '../reputation/_helpers';
 import { DEFAULT_EXPIRY_DAYS_FROM_NOW, autocomplete, execute } from './referralNew';
-import { type CreateReferralInput, createReferralCode, findExistingReferralCode } from './referralUtils';
 import { services } from './services';
+import { type CreateReferralInput, createReferralCode, findExistingReferralCode } from './utils';
 
-vi.mock('./referralUtils');
+vi.mock('./utils');
 const mockFindExistingReferralCode = vi.mocked(findExistingReferralCode);
 const mockCreateReferralCode = vi.mocked(createReferralCode);
